+++ src/main/java/org/opensearch/neuralsearch/constants/SemanticFieldConstants.java
@@ -59,4 +59,11 @@ public class SemanticFieldConstants {
      * {@link org.opensearch.neuralsearch.mapper.dto.SparseEncodingConfig}
      */
     public static final String SPARSE_ENCODING_CONFIG = "sparse_encoding_config";
+
+    /**
+     * Name of the field to configure if we should skip the embedding generation for the semantic field with existing
+     * reusable embedding. We think the embedding is reusable if it exists and the semantic field value and the
+     * model is not changed.
+     */
+    public static final String SKIP_EXISTING_EMBEDDING = "skip_existing_embedding";
 }